+++ src/components/messages/EmptyMessage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { themes } from "../styles/theme";
+import { themes } from "../../styles/theme";
 
 interface EmptyMessageProps {
   currentTheme: typeof themes.light;
