+++ src/components/messages/ErrorMessage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { themes } from "../styles/theme";
+import { themes } from "../../styles/theme";
 
 interface ErrorMessageProps {
   message: string;
