+++ src/components/messages/LoadingMessage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { themes } from "../styles/theme";
+import { themes } from "../../styles/theme";
 
 interface LoadingMessageProps {
   currentTheme: typeof themes.light;
